refactor: restructure the project into multi-module build system
- refactor into multi-module gradle build
- add spotbugs, jacoco plugins
- add plugin configs into separate gradle files
