+++ buildSrc/build.gradle
@@ -1,0 +1,7 @@
+plugins {
+  id 'groovy-gradle-plugin'
+}
+
+repositories {
+  mavenCentral()
+}
