+++ persistence/src/main/java/com/example/persistence/repositories/KnowledgeBaseRepository.java
@@ -1,0 +1,15 @@
+package com.example.persistence.repositories;
+
+import com.example.persistence.models.KnowledgeBase;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+/**
+ * KnowledgeBaseRepository.
+ *
+ * @author Bipin Thite
+ */
+@SuppressWarnings("unused")
+public interface KnowledgeBaseRepository extends JpaRepository<KnowledgeBase, Long> {
+
+  KnowledgeBase findKnowledgeBaseByName(String name);
+}
