+++ persistence/src/main/resources/application.yml
@@ -1,0 +1,46 @@
+spring:
+  jpa:
+    generate-ddl: false
+    hibernate:
+      ddl-auto: none
+    properties: 
+      hibernate:
+        dialect: org.hibernate.dialect.PostgreSQLDialect
+    show-sql: true
+
+  flyway:
+    baselineOnMigrate: true
+
+---
+
+# Spring configuration
+spring: 
+
+  config:
+    activate:
+      on-profile: default
+
+  datasource: 
+    url: jdbc:postgresql://localhost:5432/rules
+    username: rules
+    password: secretpassword
+
+---
+
+# Environment: production
+
+# Spring configuration
+spring:
+
+  config:
+    activate:
+      on-profile: production
+
+  jpa: 
+    show-sql: false
+
+  # Datasource
+  datasource: 
+    url: ${DB.URL}
+    username: ${DB.USERNAME}
+    password: ${DB.PASSWORD}
