+++ persistence/src/main/resources/db/migration/V1__20230704_initital_schema.sql
