+++ server/build.gradle
@@ -1,0 +1,7 @@
+plugins {
+  id 'rules.spring-boot'
+}
+
+dependencies {
+  implementation project(':drools')
+}
