+++ server/src/main/java/com/example/rules/Application.java
@@ -1,0 +1,17 @@
+package com.example.rules;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.context.annotation.ComponentScan;
+
+/**
+ * Spring boot application main class.
+ */
+@SuppressWarnings("PMD.UseUtilityClass")
+@ComponentScan(basePackages = "com.example")
+@SpringBootApplication
+public class Application {
+  public static void main(String[] args) {
+    SpringApplication.run(Application.class, args);
+  }
+}
