+++ server/src/main/resources/application.yml
@@ -1,0 +1,79 @@
+# Base configuration
+
+# Server configuration
+server: 
+  port: 8080
+  servlet: 
+    context-path: /api/rules
+
+spring: 
+  jpa:
+    generate-ddl: false
+    hibernate:
+      ddl-auto: none
+    properties: 
+      hibernate:
+        dialect: org.hibernate.dialect.PostgreSQLDialect
+    show-sql: true
+
+  flyway:
+    baselineOnMigrate: true
+
+logbook: 
+  include: 
+    - /api/**
+  exclude: 
+    - /actuator/**
+  write: 
+    level: INFO
+  max-body-size: 20
+  format: 
+    style: http
+  obfuscate: 
+    headers: 
+      - token
+    parameters: 
+      - email
+      - password
+
+---
+
+# Spring configuration
+spring: 
+
+  config:
+    activate:
+      on-profile: default
+
+  datasource: 
+    url: jdbc:postgresql://localhost:5432/rules
+    username: rules
+    password: secretpassword
+
+  docker:
+    compose:
+      file: "../compose.yaml"
+
+---
+
+# Environment: production
+
+# Spring configuration
+spring:
+
+  config:
+    activate:
+      on-profile: production
+
+  jpa: 
+    show-sql: false
+
+  # Datasource
+  datasource: 
+    url: ${DB.URL}
+    username: ${DB.USERNAME}
+    password: ${DB.PASSWORD}
+
+logbook: 
+  format: 
+    style: json
