--- src/main/java/com/example/rules/Application.java
@@ -1,15 +1,0 @@
-package com.example.rules;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-/**
- * Spring boot application main class.
- */
-@SuppressWarnings("PMD.UseUtilityClass")
-@SpringBootApplication
-public class Application {
-  public static void main(String[] args) {
-    SpringApplication.run(Application.class, args);
-  }
-}
