--- src/main/java/com/example/rules/repository/RuleRepository.java
@@ -1,12 +1,0 @@
-package com.example.rules.repository;
-
-import com.example.rules.model.Rule;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-/**
- * RuleRepository.
- */
-@SuppressWarnings("unused")
-public interface RuleRepository extends JpaRepository<Rule, Long> {
-  Rule findByName(String name);
-}
