--- src/main/resources/log4j2/log4j2-production.yml
@@ -1,51 +1,0 @@
-Configuration:
-  name: Default
-  properties: 
-    property: 
-    - name: appname
-      value: Rules-Engine
-
-  Appenders: 
-
-    Console: 
-      name: Console_Appender
-      target: SYSTEM_OUT
-      PatternLayout: 
-        pattern: "[%-5level] %d{yyyy-MM-dd HH:mm:ss.SSS} %c{1.} %X{requestId} - %msg%n"
-
-    GELF: 
-      name: GELF_Appender
-      server: ${env:GRAYLOG_HOST}
-      port: ${env:GRAYLOG_PORT}
-      protocol: UDP
-      PatternLayout: 
-        pattern: "%msg%n"
-      KeyValuePair: 
-        - key: facility
-          value: ${appname}
-      includeSource: false
-      includeThreadContext: true
-      includeStackTrace: false
-
-  Loggers: 
-
-      Root: 
-        level: error
-        AppenderRef: 
-          - ref: Console_Appender
- 
-      Logger: 
-        - name: com.example.rules
-          additivity: false
-          level: info
-          AppenderRef: 
-            - ref: Console_Appender
-              level: error
-            - ref: GELF_Appender
-              level: info
-
-        - name: org.zalando.logbook
-          additivity: false
-          level: info
-          AppenderRef: 
-            - ref: GELF_Appender
