--- src/main/resources/log4j2/log4j2.yml
@@ -1,34 +1,0 @@
-Configuration:
-  name: Default
-  properties: 
-    property: 
-    - name: appname
-      value: Rules-Engine
-
-  Appenders: 
-
-    Console: 
-      name: Console_Appender
-      target: SYSTEM_OUT
-      PatternLayout: 
-        pattern: "[%-5level] %d{yyyy-MM-dd HH:mm:ss.SSS} %c{1.} %X{requestId} - %msg%n"
-
-  Loggers: 
-
-      Root: 
-        level: info
-        AppenderRef: 
-          - ref: Console_Appender
-
-      Logger: 
-        - name: com.example.rules
-          additivity: false
-          level: debug
-          AppenderRef: 
-            - ref: Console_Appender
-
-        - name: org.zalando.logbook
-          additivity: false
-          level: info
-          AppenderRef: 
-            - ref: Console_Appender
